Create low-level Arrow schemas

These schemas are used as the basis for column types in Apache Arrow

Usage

geoarrow_schema_point(name = "", dim = "xy", crs = NULL, format_coord = "g")

geoarrow_schema_point_struct(
  name = "",
  dim = "xy",
  crs = NULL,
  format_coord = "g"
)

geoarrow_schema_linestring(
  name = "",
  geodesic = FALSE,
  point = geoarrow_schema_point()
)

geoarrow_schema_polygon(
  name = "",
  geodesic = FALSE,
  point = geoarrow_schema_point()
)

geoarrow_schema_multipoint(
  child,
  name = "",
  dim = "xy",
  crs = NULL,
  format_coord = "g"
)

geoarrow_schema_multilinestring(
  child,
  name = "",
  geodesic = FALSE,
  point = geoarrow_schema_point()
)

geoarrow_schema_multipolygon(
  child,
  name = "",
  geodesic = FALSE,
  point = geoarrow_schema_point()
)

geoarrow_schema_multi(child, name = "")

geoarrow_schema_wkb(name = "", format = "z", crs = NULL, geodesic = FALSE)

geoarrow_schema_wkt(name = "", format = "u", crs = NULL, geodesic = FALSE)

geoarrow_schema_geojson(name = "", format = "u", crs = NULL)

Arguments

name

An optional name

dim

A string with one character per dimension. The string must be one of xy, xyz, xym, or xyzm.

crs

A length-one character representation of the CRS. The WKT2 representation is recommended as the most complete way to encode this information; however, any string that can be recognized by the PROJ command-line utility (e.g., "OGC:CRS84").

format_coord

A format for floating point coordinate storage. This can be "f" (float/float32) or "g" (double/float64).

geodesic

Use TRUE to assert that edges should be interpolated using the shortest geodesic path (great circle on a sphere).

point

The point schema to use for coordinates

child

The child schema to use in a single-type (multi) collection

format

A custom storage format

Value

A narrow_schema().

Examples

geoarrow_schema_point()
#> <narrow_schema '+w:2' at 0x564590ada760>
#> - format: +w:2
#> - name: 
#> - flags: nullable
#> - metadata: List of 2
#>   $ ARROW:extension:name    : chr "geoarrow.point"
#>   $ ARROW:extension:metadata: raw [1:4] 00 00 00 00
#> - dictionary: NULL
#> - children[1]:
#>   <narrow_schema 'g' at 0x56458e841680>
#>   - format: g
#>   - name: xy
#>   - flags: nullable
#>   - metadata:  list()
#>   - dictionary: NULL
#>   - children[0]:
geoarrow_schema_linestring()
#> <narrow_schema '+l' at 0x564590cae4a0>
#> - format: +l
#> - name: 
#> - flags: nullable
#> - metadata: List of 2
#>   $ ARROW:extension:name    : chr "geoarrow.linestring"
#>   $ ARROW:extension:metadata: raw [1:4] 00 00 00 00
#> - dictionary: NULL
#> - children[1]:
#>   <narrow_schema '+w:2' at 0x5645911ba440>
#>   - format: +w:2
#>   - name: vertices
#>   - flags: nullable
#>   - metadata: List of 2
#>   $ ARROW:extension:name    : chr "geoarrow.point"
#>   $ ARROW:extension:metadata: raw [1:4] 00 00 00 00
#>   - dictionary: NULL
#>   - children[1]:
#>     <narrow_schema 'g' at 0x56458ea329f0>
#>     - format: g
#>     - name: xy
#>     - flags: nullable
#>     - metadata:  list()
#>     - dictionary: NULL
#>     - children[0]:
geoarrow_schema_polygon()
#> <narrow_schema '+l' at 0x564590dc7800>
#> - format: +l
#> - name: 
#> - flags: nullable
#> - metadata: List of 2
#>   $ ARROW:extension:name    : chr "geoarrow.polygon"
#>   $ ARROW:extension:metadata: raw [1:4] 00 00 00 00
#> - dictionary: NULL
#> - children[1]:
#>   <narrow_schema '+l' at 0x56459149c060>
#>   - format: +l
#>   - name: rings
#>   - flags: nullable
#>   - metadata:  list()
#>   - dictionary: NULL
#>   - children[1]:
#>     <narrow_schema '+w:2' at 0x5645912a0320>
#>     - format: +w:2
#>     - name: vertices
#>     - flags: nullable
#>     - metadata: List of 2
#>   $ ARROW:extension:name    : chr "geoarrow.point"
#>   $ ARROW:extension:metadata: raw [1:4] 00 00 00 00
#>     - dictionary: NULL
#>     - children[1]:
#>       <narrow_schema 'g' at 0x56458ea08180>
#>       - format: g
#>       - name: xy
#>       - flags: nullable
#>       - metadata:  list()
#>       - dictionary: NULL
#>       - children[0]:
geoarrow_schema_multi(geoarrow_schema_point())
#> <narrow_schema '+l' at 0x564590dc7800>
#> - format: +l
#> - name: 
#> - flags: nullable
#> - metadata: List of 2
#>   $ ARROW:extension:name    : chr "geoarrow.multi"
#>   $ ARROW:extension:metadata: raw [1:4] 00 00 00 00
#> - dictionary: NULL
#> - children[1]:
#>   <narrow_schema '+w:2' at 0x56459149c060>
#>   - format: +w:2
#>   - name: points
#>   - flags: nullable
#>   - metadata: List of 2
#>   $ ARROW:extension:name    : chr "geoarrow.point"
#>   $ ARROW:extension:metadata: raw [1:4] 00 00 00 00
#>   - dictionary: NULL
#>   - children[1]:
#>     <narrow_schema 'g' at 0x564591168d60>
#>     - format: g
#>     - name: xy
#>     - flags: nullable
#>     - metadata:  list()
#>     - dictionary: NULL
#>     - children[0]: